requirements.txt
Flask==3.1.1
flask-cors==5.0.0
mysql-connector-python==9.5.0
